reorganized media control buttons to own files for easier unit testing
