+++ client/src/components/MediaControlsBar/MediaControls/RepeatButton/RepeatButton.css
@@ -1,0 +1,12 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer components {
+  .repeating {
+    @apply fill-green-450
+      hover:fill-green-500
+      dark:fill-green-350 
+      hover:dark:fill-green-450;
+  }
+}
